requirements-archive.txt
# -- requirements.txt -----

flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
flask-mysqldb>=1.0.1
Flask-Mail>=0.9.1
